Move `crate_target_triple`.
